Allowing deepcopying Constant objects
This allows deepcopying Django model instances. This has become relevant
since Django 3.2 started making deepcopies of attributes set in the
setUpTestData() method of testcases.

This adds a __deepcopy__() method to make the deepcopy. Without this,
Python will make a copy itself, but it seems that might end up querying
other attributes before the constants attribute is assigned, leading to:

    `constants` not set on Constant so cannot check equality